chore: refactor guardian package
