+++ packages/guardian/src/services/index.ts
@@ -1,5 +1,5 @@
 export * from "./GuardianSigner"
-export * from "./account"
+export * from "./CosignerTypes"
 export * from "./verifiedEmail"
 export * from "./idb"
 export * from "./jwt"
